chore: rename files to match class
